[#1]NEW : Authentication feature
Fixes #1
This feature give capability to user to create an account and logging in
to Tournament Factory.
By default, all user are logged in ROLE_USER.
A firewall has been installed and parameterized for control route access
This commit is comming with a new migration for DB, TFUser Entity has been
added.